refac: value from double to numeric

+++ models/procedure.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, DateTime, Double, Enum, ForeignKey, Integer
+from sqlalchemy import Column, DateTime, Enum, ForeignKey, Integer, Numeric
 from sqlalchemy.orm import relationship
 
 from .base import Base
@@ -11,7 +11,7 @@ class Procedure(Base):
     doctor_id = Column(Integer, ForeignKey("doctors.id"), nullable=False)
     patient_id = Column(Integer, ForeignKey("patients.id"), nullable=False)
     date = Column(DateTime, nullable=False)
-    value = Column(Double, nullable=False)
+    value = Column(Numeric(10, 2), nullable=False)
     payment_status = Column(
         Enum("paid", "pending", "glossed", name="payment_status"), nullable=False
     )
